Fix timezone in serialized dates
BREAKING CHANGE: this affects the server-side parsing of dates sent by
this library. Previously, the server had to completely ignore the
timezone information ("UTC"), then treat the date as local. Now the
front-end will properly send the local date plus the local timezone,
which allows the back-end to reliably choose between either handling it
as a local time, or convert it to some other timezone, like UTC.