Update model when server responds with `200 OK`
When the server responds with the status `200 OK` the client should check, if the server updated the resource as well and provides an updated version of the resource. In that case the `data` attribute is populated with the updated resource. When the server didn't change anything there is no `data` attribute and we can return the local model.

See http://jsonapi.org/format/#crud-updating